requirements.txt
speedtest-cli==2.1.3
colorama==0.4.6
tqdm==4.66.1